+++ lizmap_server/__init__.py
@@ -19,7 +19,14 @@ def classFactory(iface):
             QMessageBox.warning(
                 self.iface.mainWindow(),
                 'Lizmap server plugin',
-                'Lizmap server is plugin for QGIS Server. For QGIS Desktop, use the other plugin called "Lizmap".',
+                '"Lizmap server" is a plugin designed <b>only</b> for QGIS Server.'
+                '<br><br>'
+                'This plugin must be installed on QGIS Server following the documentation about '
+                '<a href="https://docs.qgis.org/latest/en/docs/server_manual/plugins.html">'
+                'QGIS server side plugins'
+                '</a>.'
+                '<br><br>'
+                'For QGIS Desktop, use the other plugin called "Lizmap".',
             )
 
         def unload(self):